requirements.txt
websockets>=13.0
sounddevice>=0.5.0
python-dotenv>=1.0.0
